Chapter 10 Lab
This program uses Exceptions and Advanced file I/O. In this program it takes two static arguments, a file name and an int array that opens the content of a binary file and writes to it an the closes it. Then a second static method reads to a file. and the data from the file is printed to the screen.